Updated Image Scraper
imported a traceback library to assist with error messages.
Image Scrapper function is now working somewhat correctly. It scraps images and correctly assigns them a name in a folder, however these images seem to get corrupted in a way that makes them unable to be opened.

Still need to do the text scraper method properly.