Tiny text_scrape update
added clarity comments to help myself. Now the method creates and names the .txt file. It was writing but a bug in the for loop caused it to repeatedly re-write itself, its also not writing the information I wanted so i've moved some stuff around and will approach the problem again in a new way.